Use unsafe get function for getting slices
The MMU does all of the bounds checking manually to make sure it will
catch any improper access to the backing memory. This means that we
can safely use the unsafe `get_unchecked*` functions to get the slice
of data that is being accessed.

This gives a slight performance benefit to some operations.