Use box slice instead of vec in map
Using the boxed slice should decrease the overall size
of the map struct. That will improve cache locality and
overall performance.

From running the benchmarks, it appears that it helps
with resetting the memory when changes have been
made but makes the performance slightly worse when
no changes have been made.

I think that is a win since changes will generally be made.