Don't implement from Perm for Reason
Changes the From implementation to be a crate private method
on the enum instead of being public. Not any Perm can actually be converted into a Reason and some of them will cause a panic. Because of that, it's not truly a From implementation and that
implementation was not really meant for use outside of the
crate anyways. Making it a private method makes the API clearer
and more intuitive to use.